WIP: correctly scale magnetometer data to mG
Need output in milligauss, but the code from ChibiOS
was multiplying the already scaled driver output in gauss
by 1000 and then dividing by 4096.

+++ src/magnetometer.h
@@ -4,7 +4,7 @@
 //#define DEV_MAG_ZEPHYR_ENABLE_MAGB
 #define NUM_MAGS 2
 
-int get_mag_reading(int mag_num, int16_t *x, int16_t *y, int16_t *z);
+int get_mag_reading(int mag_num, int32_t *x, int32_t *y, int32_t *z);
 
 #endif
